content/media/webaudio/test/test_oscillatorNodeStart.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test the OscillatorNode interface</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {

  var context = new AudioContext();
  var osc = context.createOscillator();
  var sp = context.createScriptProcessor();

  osc.connect(sp);

  sp.onaudioprocess = function (e) {
    var input = e.inputBuffer.getChannelData(0);
    var isSilent = true;
    for (var i = 0; i < input.length; i++) {
      if (input[i] != 0.0) {
        isSilent = false;
      }
    }
    sp.onaudioprocess = null;
    dump("audioprocess received\n");
    ok(isSilent, "OscillatorNode should be silent before calling start.");
    SimpleTest.finish();
  }

  // Debug Logging for bug 966322

  dump("Load event ran\n");

  var bs = context.createBufferSource();
  bs.buffer = context.createBuffer(1, 1, context.sampleRate);
  bs.start();
  bs.onended = function() {
    dump("Graph is running\n");
  }

});

// Debug Logging for bug 966322
dump("script ran\n");

</script>
</pre>
</body>
</html>
